Restore i2c_ssd1306_legacy example, remove ungated copy
